ci: extend web response-surface check to cover full nginx header set
Rename web-csp-render-check.mjs to web-response-surface-check.mjs and
extend it to browser-verify X-Content-Type-Options, X-Frame-Options,
Referrer-Policy, Cache-Control/Pragma/Expires, and CSP+nonce presence
against the real navigation response, alongside the existing CSP/CSS
render assertions. X-Frame-Options falls back to a mechanism-level
header check with a documented rationale, after empirically confirming
Chromium does not surface its enforcement through any observable
page-level signal. hashed/static-asset caching and server_tokens are
explicitly excluded as out of scope for a single-route check.
