+++ docs/KNOWLEDGE_INBOX.md
@@ -20,7 +20,7 @@ Why: an implementation agent deleting the now-empty `libs/budget/validation` lib
 
 ## 2026-08-08 — [devops] `nginx.conf`'s `sub_filter` silently no-ops on zero matches — exact-string substitutions (e.g. CSP nonce injection into built `index.html`) need an explicit smoke check, not just a green build
 
-Why: `apps/web/nginx.conf` injects a per-request CSP nonce into two exact-string targets in Angular's built `index.html` via `sub_filter`. `sub_filter` has no error/warning path for zero matches — if a future Angular/build-tool upgrade changes attribute order or quoting, the substitution silently stops applying and the CSP nonce mismatch bug reappears with no CI signal. This already happened once and was only caught by manual browser inspection. **Update 2026-08-09: the first remedy — `scripts/web-csp-smoke-check.sh` grepping the served `index.html` for both nonce patterns — was itself insufficient and passed green through a completely unstyled production app.** Verifying that a substitution _landed_ is not the same as verifying the page _works_; see the 2026-08-09 `[testing]` entry on mechanism-vs-outcome assertions. The check now drives headless Chromium (`scripts/web-csp-render-check.mjs`) and the greps survive only as diagnostics. The container-start mechanics are still accurate and still useful: `docker run ... --add-host api:127.0.0.1`, needed so nginx can resolve its upstream without the full compose network, then wait on `/health`. This entry was also re-confirmed the hard way — disabling `inlineCritical` changed the meta from `content>` to `content=""` and would have silently no-opped the remaining `sub_filter`. Generalises: any nginx `sub_filter`/`proxy_set_header`-style exact-string config substitution should carry a build-time or deploy-time assertion on the substituted output, since nginx itself won't flag a silent no-op. Belongs in (guess): `rules/local/architecture-backend.md` or a new `rules/local/nginx.md` if nginx config recipes accumulate — ambiguous because this is currently the only nginx-specific gotcha recorded in this repo.
+Why: `apps/web/nginx.conf` injects a per-request CSP nonce into two exact-string targets in Angular's built `index.html` via `sub_filter`. `sub_filter` has no error/warning path for zero matches — if a future Angular/build-tool upgrade changes attribute order or quoting, the substitution silently stops applying and the CSP nonce mismatch bug reappears with no CI signal. This already happened once and was only caught by manual browser inspection. **Update 2026-08-09: the first remedy — `scripts/web-csp-smoke-check.sh` grepping the served `index.html` for both nonce patterns — was itself insufficient and passed green through a completely unstyled production app.** Verifying that a substitution _landed_ is not the same as verifying the page _works_; see the 2026-08-09 `[testing]` entry on mechanism-vs-outcome assertions. The check now drives headless Chromium (`scripts/web-response-surface-check.mjs`) and the greps survive only as diagnostics. The container-start mechanics are still accurate and still useful: `docker run ... --add-host api:127.0.0.1`, needed so nginx can resolve its upstream without the full compose network, then wait on `/health`. This entry was also re-confirmed the hard way — disabling `inlineCritical` changed the meta from `content>` to `content=""` and would have silently no-opped the remaining `sub_filter`. Generalises: any nginx `sub_filter`/`proxy_set_header`-style exact-string config substitution should carry a build-time or deploy-time assertion on the substituted output, since nginx itself won't flag a silent no-op. Belongs in (guess): `rules/local/architecture-backend.md` or a new `rules/local/nginx.md` if nginx config recipes accumulate — ambiguous because this is currently the only nginx-specific gotcha recorded in this repo.
 
 ## 2026-08-08 — [shell] Two `sed`-in-Dockerfile pitfalls that each silently produce a wrong-but-plausible result, caught only by independently recomputing the expected value
 
@@ -101,3 +101,7 @@ Why: a drafted `.github/dependabot.yml` included a `reviewers:` block under an `
 ## 2026-08-09 — [docs-tooling] Distillation must check the target section for an existing near-duplicate before writing — the inbox's own dedup warning only covers the parallel changelog, not `rules/local/**`
 
 Why: distilling two entries, the writing agent found that `rules/local/testing.md` already carried "A green, clearly-named test can assert a value that violates a contract enforced elsewhere" (a near-duplicate of the weak-assertion entry it was inlining) and `rules/local/code-style-angular.md` already carried an "Error Object Translations" section stating the same principle as the translation-key entry. Both were partial distillations from an earlier session under different framing. The agent merged into the existing sections rather than appending parallel ones, which is the right outcome — but nothing in the process required that check. This inbox's header warns to check `docs/CLAUDE_TS_CHANGELOG.md` for an existing richer record before re-adding an entry; it does not warn that the destination rule file may already state the same thing. Without the check, repeated distillation runs accrete near-duplicate sections that slowly disagree with each other. Belongs in: `.claude/skills/distill-inbox/SKILL.md` (a pre-write step in Step 4's dispatch template, and a Step 4.5 verification item).
+
+## 2026-08-09 — [testing] Chromium does not surface `X-Frame-Options` frame-blocking via `page.on('console')`, and `contentDocument`/frame-load-timing signals cannot distinguish "blocked by X-Frame-Options" from "loaded but same-origin-policy-restricted" — outcome-level browser testing of this header is not achievable the way CSP-violation testing is
+
+Why: extending `scripts/web-response-surface-check.mjs` to browser-verify `X-Frame-Options: SAMEORIGIN` (previously only asserted via `nginx.conf` text), the natural approach — by analogy with the file's existing CSP-violation check, which does work via `page.on('console')` — was to embed the served page in an iframe from a cross-origin (`data:`) parent and watch for the browser's refusal signal. This was tried three times before being empirically confirmed to fail: against the _correct_ config (header present, framing should be refused), no console message naming an X-Frame-Options refusal ever appeared, the iframe's `onload` fired normally, and `contentDocument` access was denied — but that denial is indistinguishable from ordinary same-origin-policy restriction on a `data:`-origin parent reading a _successfully loaded_ cross-origin child, which would look identical whether or not the header did anything. No page-level Playwright signal reliably distinguishes "frame blocked by X-Frame-Options" from "frame loaded fine but is opaque to JS for unrelated reasons" in this environment. The correct fallback, per the task's own documented escape hatch, was a mechanism-level header check with an inline comment recording what was tried and why it failed — not another attempt at the same technique. Generalizes: before assuming a CSP-violation-style browser-observation pattern transfers to a different security header, verify the browser actually surfaces that header's enforcement through an observable channel (console, network, DOM) — CSP does; X-Frame-Options, empirically, does not. Belongs in: `rules/local/testing.md`, as a browser-render-check pattern alongside the existing mechanism-vs-outcome principle.
